backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
django-mongodb-backend>=5.1
pymongo>=4.6
